conclui backend

+++ backend/src/main/java/com/laercio/desafio_fast/workshop/repository/ColaboradorRepository.java
@@ -3,6 +3,6 @@ package com.laercio.desafio_fast.workshop.repository;
 import com.laercio.desafio_fast.workshop.model.Colaborador;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ColaboradorRepository extends JpaRepository<Colaborador, Integer> {
+public interface ColaboradorRepository extends JpaRepository<Colaborador, Long> {
 }
